pyproject.toml
[tool.poetry]
name = "color-palette"
version = "0.1.0"
description = ""
authors = ["Dmitrii Turetskikh <tureckihdmitriq@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
passlib = "^1.7.4"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
bcrypt = "^4.1.2"
python-multipart = "^0.0.9"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
